chore: enable all breaking changes
